adding system property to specify further smt-solvers.json
