--- key.core/src/main/resources/de/uka/ilkd/key/smt/solvertypes/CVC4.props
@@ -1,23 +1,0 @@
-experimental=true
-params=--no-print-success --produce-models --no-interactive --lang smt2
-timeout=-1
-name=CVC4
-info=CVC4 solver (https://cvc4.github.io/index.html).
-command=cvc4
-version=--version
-socketClass=de.uka.ilkd.key.smt.communication.CVC4Socket
-handlers=de.uka.ilkd.key.smt.newsmt2.BooleanConnectiveHandler,\
-de.uka.ilkd.key.smt.newsmt2.PolymorphicHandler,\
-de.uka.ilkd.key.smt.newsmt2.QuantifierHandler,\
-de.uka.ilkd.key.smt.newsmt2.LogicalVariableHandler,\
-de.uka.ilkd.key.smt.newsmt2.NumberConstantsHandler,\
-de.uka.ilkd.key.smt.newsmt2.IntegerOpHandler,\
-de.uka.ilkd.key.smt.newsmt2.InstanceOfHandler,\
-de.uka.ilkd.key.smt.newsmt2.CastHandler,\
-de.uka.ilkd.key.smt.newsmt2.SumProdHandler,\
-de.uka.ilkd.key.smt.newsmt2.UpdateHandler,\
-de.uka.ilkd.key.smt.newsmt2.FieldConstantHandler,\
-de.uka.ilkd.key.smt.newsmt2.FloatHandler,\
-de.uka.ilkd.key.smt.newsmt2.CastingFunctionsHandler,\
-de.uka.ilkd.key.smt.newsmt2.DefinedSymbolsHandler,\
-de.uka.ilkd.key.smt.newsmt2.UninterpretedSymbolsHandler
