--- key.core/src/main/resources/de/uka/ilkd/key/smt/solvertypes/CVC4_legacy.props
@@ -1,9 +1,0 @@
-experimental=true
-params=--no-print-success --produce-models --no-interactive --lang smt2
-timeout=-1
-name=CVC4 (Legacy Translation)
-info=CVC4 solver
-command=cvc4
-version=--version
-socketClass=de.uka.ilkd.key.smt.communication.CVC4Socket
-translatorClass=de.uka.ilkd.key.smt.SmtLib2Translator
