--- key.core/src/main/resources/de/uka/ilkd/key/smt/solvertypes/CVC5.props
@@ -1,23 +1,0 @@
-params=--no-interactive --produce-models --lang smt2
-timeout=-1
-name=cvc5
-info=cvc5 solver (https://cvc5.github.io/).
-command=cvc5
-version=--version
-socketClass=de.uka.ilkd.key.smt.communication.Cvc5Socket
-handlers=de.uka.ilkd.key.smt.newsmt2.BooleanConnectiveHandler,\
-de.uka.ilkd.key.smt.newsmt2.PolymorphicHandler,\
-de.uka.ilkd.key.smt.newsmt2.QuantifierHandler,\
-de.uka.ilkd.key.smt.newsmt2.LogicalVariableHandler,\
-de.uka.ilkd.key.smt.newsmt2.NumberConstantsHandler,\
-de.uka.ilkd.key.smt.newsmt2.IntegerOpHandler,\
-de.uka.ilkd.key.smt.newsmt2.InstanceOfHandler,\
-de.uka.ilkd.key.smt.newsmt2.CastHandler,\
-de.uka.ilkd.key.smt.newsmt2.SumProdHandler,\
-de.uka.ilkd.key.smt.newsmt2.UpdateHandler,\
-de.uka.ilkd.key.smt.newsmt2.FieldConstantHandler,\
-de.uka.ilkd.key.smt.newsmt2.FloatHandler,\
-de.uka.ilkd.key.smt.newsmt2.CastingFunctionsHandler,\
-de.uka.ilkd.key.smt.newsmt2.DefinedSymbolsHandler,\
-de.uka.ilkd.key.smt.newsmt2.UninterpretedSymbolsHandler
-handlerOptions=getUnsatCore
