+++ key.core/src/test/java/de/uka/ilkd/key/smt/TestSMTFocusResults.java
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: GPL-2.0-only */
 package de.uka.ilkd.key.smt;
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 /**
  * Tests for {@link SMTFocusResults}.
@@ -13,8 +14,8 @@ class TestSMTFocusResults {
     @Test
     void parsesZ3Format() {
         String z3Output = "(L_1 L_17 L_23)";
-        Integer[] result = SMTFocusResults.parseZ3Format(z3Output);
-        Assertions.assertArrayEquals(new Integer[] { 1, 17, 23 }, result);
+        var result = SMTFocusResults.parseZ3Format(z3Output);
+        assertThat(result).containsOnly(1, 17, 23);
     }
 
     @Test
@@ -26,7 +27,7 @@ class TestSMTFocusResults {
             "L_23",
             ")"
         };
-        Integer[] result = SMTFocusResults.parseCVC5Format(cvc5Output);
-        Assertions.assertArrayEquals(new Integer[] { 5, 12, 23 }, result);
+        var result = SMTFocusResults.parseCVC5Format(cvc5Output);
+        assertThat(result).containsOnly(5, 12, 23);
     }
 }
