+++ key.core/src/test/java/de/uka/ilkd/key/smt/TestUnsatCore.java
@@ -4,6 +4,7 @@
 package de.uka.ilkd.key.smt;
 
 import java.nio.file.Path;
+import java.util.Objects;
 
 import de.uka.ilkd.key.control.DefaultUserInterfaceControl;
 import de.uka.ilkd.key.control.KeYEnvironment;
@@ -16,14 +17,17 @@ import org.key_project.prover.sequent.PosInOccurrence;
 import org.key_project.util.collection.ImmutableList;
 import org.key_project.util.helper.FindResources;
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
 
 /**
  * Tests for the unsat core saving infrastructure.
  */
 class TestUnsatCore {
-    private static final Path testCaseDirectory = FindResources.getTestCasesDirectory();
+    private static final Path testCaseDirectory =
+        Objects.requireNonNull(FindResources.getTestCasesDirectory());
 
     @Test
     void testUnsatCore() throws ProblemLoaderException {
@@ -31,26 +35,23 @@ class TestUnsatCore {
 
         KeYEnvironment<DefaultUserInterfaceControl> env =
             KeYEnvironment.load(testCaseDirectory.resolve("smt/unsatCore.proof"));
-        Assertions.assertNotNull(env.getLoadedProof());
-        Assertions.assertTrue(env.getLoadedProof().closed());
+        assertNotNull(env.getLoadedProof());
+        assertTrue(env.getLoadedProof().closed());
         // find the SMT rule app
         Proof p = env.getLoadedProof();
         Node n = p.findAny(node -> node.getAppliedRuleApp() instanceof SMTRuleApp);
+        assertNotNull(n);
         SMTRuleApp app = ((SMTRuleApp) n.getAppliedRuleApp());
-        Assertions.assertEquals("Z3", app.getSuccessfulSolverName());
+        System.out.println(app);
+        assertEquals("Z3", app.getSuccessfulSolverName());
         ImmutableList<PosInOccurrence> ifs = app.assumesInsts();
-        Assertions.assertTrue(
-            ifs.contains(PosInOccurrence.findInSequent(n.sequent(), 1,
-                PosInTerm.getTopLevel())));
-        Assertions.assertTrue(
-            ifs.contains(PosInOccurrence.findInSequent(n.sequent(), 2,
-                PosInTerm.getTopLevel())));
-        Assertions.assertTrue(
-            ifs.contains(PosInOccurrence.findInSequent(n.sequent(), 3,
-                PosInTerm.getTopLevel())));
-        Assertions.assertTrue(
-            ifs.contains(PosInOccurrence.findInSequent(n.sequent(), 7, PosInTerm.getTopLevel())));
-        Assertions.assertEquals(4, ifs.size());
+        var pio1 = PosInOccurrence.findInSequent(n.sequent(), 1, PosInTerm.getTopLevel());
+        var pio2 = PosInOccurrence.findInSequent(n.sequent(), 2, PosInTerm.getTopLevel());
+        var pio3 = PosInOccurrence.findInSequent(n.sequent(), 3, PosInTerm.getTopLevel());
+        var pio7 = PosInOccurrence.findInSequent(n.sequent(), 7, PosInTerm.getTopLevel());
+
+        assertEquals(4, ifs.size());
+        assertThat(ifs).containsExactlyInAnyOrder(pio1, pio2, pio3, pio7);
     }
 
 }
